fix lon/lat parsing
git-svn-id: http://buddycloud.googlecode.com/svn/trunk/client/android/main@110 3fc15f3e-b8c3-11de-ba48-051263943efc

+++ src/com/buddycloud/jbuddycloud/packet/LocationEvent.java
@@ -14,9 +14,9 @@ public class LocationEvent implements PacketExtension {
 	public String locality;
 	public String region;
 	public String country;
-	public int lat;
-	public int lng;
-	public int accuracy;
+	public double lat;
+	public double lng;
+	public double accuracy;
 	public int type;
 
 	public String getElementName() {
